Package rename for java_type.
