+++ pkg/common/encoding/java_type/java_type.go
@@ -1,4 +1,4 @@
-package javaType
+package java_type
 
 /*
  * Licensed to the Apache Software Foundation (ASF) under one or more
